% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psylisten.R
\docType{package}
\name{psylisten-package}
\alias{psylisten}
\alias{psylisten-package}
\title{psylisten: Collection of LongGold psychTestR listening tests}
\description{
Implementation of a LongGold test battery consisting of a selection of listening tests based on the psychTestR library. Demos of the whole set of performance tests, listening tests, and questionnaires can be accessed at https://shiny.gold-msi.org/longgold_demo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmhoeger/psylisten}
  \item Report bugs at \url{https://github.com/fmhoeger/psylisten/issues}
}

}
\keyword{internal}
