% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{psylisten_deps}
\alias{psylisten_deps}
\title{List all psylisten dependencies}
\usage{
psylisten_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
psylisten packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
List all psylisten dependencies
}
