% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psylisten_packages}
\alias{psylisten_packages}
\title{List all packages in the psylisten}
\usage{
psylisten_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include psylisten in the list?}
}
\description{
List all packages in the psylisten
}
\examples{
psylisten_packages()
}
